+++ src/Symfony/Component/VarDumper/Tests/Caster/ResourceCasterTest.php
@@ -12,82 +12,49 @@
 namespace Symfony\Component\VarDumper\Tests\Caster;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
+use Symfony\Component\VarDumper\Caster\ResourceCaster;
+use Symfony\Component\VarDumper\Cloner\Stub;
 use Symfony\Component\VarDumper\Test\VarDumperTestTrait;
 
 class ResourceCasterTest extends TestCase
 {
+    use ExpectDeprecationTrait;
     use VarDumperTestTrait;
 
-    public function testCastCurl()
+    /**
+     * @group legacy
+     *
+     * @requires extension curl
+     */
+    public function testCastCurlIsDeprecated()
     {
         $ch = curl_init('http://example.com');
         curl_setopt($ch, \CURLOPT_RETURNTRANSFER, true);
         curl_exec($ch);
 
-        $this->assertDumpMatchesFormat(
-            <<<'EODUMP'
-CurlHandle {
-  url: "http://example.com/"
-  content_type: "text/html; charset=UTF-8"
-  http_code: 200%A
-}
-EODUMP, $ch);
+        $this->expectDeprecation('Since symfony/var-dumper 7.3: The "Symfony\Component\VarDumper\Caster\ResourceCaster::castCurl()" method is deprecated without replacement.');
+
+        ResourceCaster::castCurl($ch, [], new Stub(), false);
     }
 
-    public function testAsymmetricKey()
+    /**
+     * @group legacy
+     *
+     * @requires extension gd
+     */
+    public function testCastGdIsDeprecated()
     {
-        $key = openssl_pkey_new([
-            'private_key_bits' => 1024,
-            'private_key_type' => \OPENSSL_KEYTYPE_RSA,
-        ]);
+        $gd = imagecreate(1, 1);
 
-        $this->assertDumpMatchesFormat(
-            <<<'EODUMP'
-OpenSSLAsymmetricKey {
-  bits: 1024
-  key: """
-    -----BEGIN PUBLIC KEY-----\n
-    %A
-    %A
-    %A
-    %A
-    -----END PUBLIC KEY-----\n
-    """
-  type: 0
-}
-EODUMP, $key);
-    }
+        $this->expectDeprecation('Since symfony/var-dumper 7.3: The "Symfony\Component\VarDumper\Caster\ResourceCaster::castGd()" method is deprecated without replacement.');
 
-    public function testOpensslCsr()
-    {
-        $dn = [
-            'countryName' => 'FR',
-            'stateOrProvinceName' => 'Ile-de-France',
-            'localityName' => 'Paris',
-            'organizationName' => 'Symfony',
-            'organizationalUnitName' => 'Security',
-            'commonName' => 'symfony.com',
-            'emailAddress' => 'test@symfony.com',
-        ];
-        $privkey = openssl_pkey_new();
-        $csr = openssl_csr_new($dn, $privkey);
-
-        $this->assertDumpMatchesFormat(
-            <<<'EODUMP'
-OpenSSLCertificateSigningRequest {
-  countryName: "FR"
-  stateOrProvinceName: "Ile-de-France"
-  localityName: "Paris"
-  organizationName: "Symfony"
-  organizationalUnitName: "Security"
-  commonName: "symfony.com"
-  emailAddress: "test@symfony.com"
-}
-EODUMP, $csr);
+        ResourceCaster::castGd($gd, [], new Stub(), false);
     }
 
     /**
      * @requires PHP < 8.4
+     * @requires extension dba
      */
     public function testCastDbaPriorToPhp84()
     {
@@ -97,6 +64,44 @@ EODUMP, $csr);
             <<<'EODUMP'
 dba resource {
   file: %s
+}
+EODUMP, $dba);
+    }
+
+    /**
+     * @requires PHP 8.4
+     */
+    public function testCastDba()
+    {
+        if (\PHP_VERSION_ID >= 80400 && \PHP_VERSION_ID < 80402) {
+            $this->markTestSkipped('The test cannot be run on PHP 8.4.0 and PHP 8.4.1, see https://github.com/php/php-src/issues/16990');
+        }
+
+        $dba = dba_open(sys_get_temp_dir().'/test.db', 'c');
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Dba\Connection {
+  +file: %s
+}
+EODUMP, $dba);
+    }
+
+
+    /**
+     * @requires PHP 8.4
+     */
+    public function testCastDbaOnBuggyPhp84()
+    {
+        if (\PHP_VERSION_ID >= 80402) {
+            $this->markTestSkipped('The test can only be run on PHP 8.4.0 and 8.4.1, see https://github.com/php/php-src/issues/16990');
+        }
+
+        $dba = dba_open(sys_get_temp_dir().'/test.db', 'c');
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+dba resource {
 }
 EODUMP, $dba);
     }