+++ src/Symfony/Component/VarDumper/Tests/Caster/SocketCasterTest.php
@@ -21,6 +21,9 @@ class SocketCasterTest extends TestCase
 {
     use VarDumperTestTrait;
 
+    /**
+     * @requires PHP 8.3
+     */
     public function testCastSocket()
     {
         $socket = socket_create(\AF_INET, \SOCK_DGRAM, \SOL_UDP);
@@ -29,15 +32,100 @@ class SocketCasterTest extends TestCase
         $this->assertDumpMatchesFormat(
             <<<'EODUMP'
 Socket {
-  address: "127.0.0.1"
-  port: %d
+  uri: "udp://127.0.0.1:%d"
+  timed_out: false
+  blocked: true%A
+}
+EODUMP, $socket);
+    }
+
+    /**
+     * @requires PHP < 8.3
+     */
+    public function testCastSocketPriorToPhp83()
+    {
+        $socket = socket_create(\AF_INET, \SOCK_DGRAM, \SOL_UDP);
+        @socket_connect($socket, '127.0.0.1', 80);
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Socket {
   timed_out: false
   blocked: true
-  eof: false
-  stream_type: "udp_socket"
-  mode: "r+"
-  unread_bytes: 0
-  seekable: false%A
+}
+EODUMP, $socket);
+    }
+
+    /**
+     * @requires PHP 8.3
+     */
+    public function testCastSocketIpV6()
+    {
+        $socket = socket_create(\AF_INET6, \SOCK_STREAM, \SOL_TCP);
+        @socket_connect($socket, '::1', 80);
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Socket {
+  uri: "tcp://[%A]:%d"
+  timed_out: false
+  blocked: true
+  last_error: SOCKET_ECONNREFUSED
+}
+EODUMP, $socket);
+    }
+
+    /**
+     * @requires PHP < 8.3
+     */
+    public function testCastSocketIpV6PriorToPhp83()
+    {
+        $socket = socket_create(\AF_INET6, \SOCK_STREAM, \SOL_TCP);
+        @socket_connect($socket, '::1', 80);
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Socket {
+  timed_out: false
+  blocked: true
+  last_error: SOCKET_ECONNREFUSED
+}
+EODUMP, $socket);
+    }
+
+    /**
+     * @requires PHP 8.3
+     */
+    public function testCastUnixSocket()
+    {
+        $socket = socket_create(\AF_UNIX, \SOCK_STREAM, 0);
+        @socket_connect($socket, '/tmp/socket.sock');
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Socket {
+  uri: "unix://"
+  timed_out: false
+  blocked: true
+  last_error: SOCKET_ENOENT
+}
+EODUMP, $socket);
+    }
+
+    /**
+     * @requires PHP < 8.3
+     */
+    public function testCastUnixSocketPriorToPhp83()
+    {
+        $socket = socket_create(\AF_UNIX, \SOCK_STREAM, 0);
+        @socket_connect($socket, '/tmp/socket.sock');
+
+        $this->assertDumpMatchesFormat(
+            <<<'EODUMP'
+Socket {
+  timed_out: false
+  blocked: true
+  last_error: SOCKET_ENOENT
 }
 EODUMP, $socket);
     }